MLP Clean Architecture
