--- ClassLibrary1/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace ClassLibrary1
-{
-    public class Class1
-    {
-
-    }
-}
