+++ ClassLibrary1/ClassLibrary1.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <None Remove="ClassLibrary1.v3.ncrunchproject" />
+  </ItemGroup>
+
 </Project>
