+++ ClassLibrary1/ClassLibrary1.v3.ncrunchproject
@@ -1,0 +1,5 @@
+﻿<ProjectConfiguration>
+  <Settings>
+    <HiddenComponentWarnings />
+  </Settings>
+</ProjectConfiguration>
