+++ ClassLibrary1/SiteMap.cs
@@ -1,0 +1,8 @@
+﻿namespace ClassLibrary1
+{
+    public static class EndpointMap
+    {
+        public const string IoEndpoint = "/io";
+        public const string HealthCheckEndpoint = "/health";
+    }
+}
