+++ TestProject1/AuthenticationConfig.cs
@@ -1,0 +1,102 @@
+﻿using Microsoft.Extensions.Configuration;
+using System;
+using System.Globalization;
+using System.IO;
+
+namespace TestProject1
+{
+    /// <summary>
+    /// Description of the configuration of an AzureAD public client application (desktop/mobile application). This should
+    /// match the application registration done in the Azure portal
+    /// </summary>
+    public class AuthenticationConfig
+    {
+        private string tenant = "";
+        private string clientId = "";
+        private string clientSecret = "";
+        private string certificateName = "";
+        private string todoListBaseAddress = "";
+        private string todoListScope = "";
+
+        /// <summary>
+        /// instance of Azure AD, for example public Azure or a Sovereign cloud (Azure China, Germany, US government, etc ...)
+        /// </summary>
+        public string Instance { get; set; } = "https://login.microsoftonline.com/{0}";
+
+        /// <summary>
+        /// The Tenant is:
+        /// - either the tenant ID of the Azure AD tenant in which this application is registered (a guid)
+        /// or a domain name associated with the tenant
+        /// - or 'organizations' (for a multi-tenant application)
+        /// </summary>
+        public string Tenant { get => tenant; set => tenant = value; }
+
+        /// <summary>
+        /// Guid used by the application to uniquely identify itself to Azure AD
+        /// </summary>
+        public string ClientId { get => clientId; set => clientId = value; }
+
+        /// <summary>
+        /// URL of the authority
+        /// </summary>
+        public string Authority
+        {
+            get
+            {
+                return String.Format(CultureInfo.InvariantCulture, Instance, Tenant);
+            }
+        }
+
+        /// <summary>
+        /// Client secret (application password)
+        /// </summary>
+        /// <remarks>Daemon applications can authenticate with AAD through two mechanisms: ClientSecret
+        /// (which is a kind of application password: this property)
+        /// or a certificate previously shared with AzureAD during the application registration 
+        /// (and identified by the CertificateName property belows)
+        /// <remarks> 
+        public string ClientSecret { get => clientSecret; set => clientSecret = value; }
+
+        /// <summary>
+        /// Name of a certificate in the user certificate store
+        /// </summary>
+        /// <remarks>Daemon applications can authenticate with AAD through two mechanisms: ClientSecret
+        /// (which is a kind of application password: the property above)
+        /// or a certificate previously shared with AzureAD during the application registration 
+        /// (and identified by this CertificateName property)
+        /// <remarks> 
+        public string CertificateName { get => certificateName; set => certificateName = value; }
+
+        /// <summary>
+        /// Web Api base URL
+        /// </summary>
+        public string TodoListBaseAddress { get => todoListBaseAddress; set => todoListBaseAddress = value; }
+
+        /// <summary>
+        /// Web Api scope. With client credentials flows, the scopes is ALWAYS of the shape "resource/.default"
+        /// </summary>
+        public string TodoListScope { get => todoListScope; set => todoListScope = value; }
+
+        /// <summary>
+        /// Reads the configuration from a json file
+        /// </summary>
+        /// <param name="path">Path to the configuration json file</param>
+        /// <returns>AuthenticationConfig read from the json file</returns>
+        public static AuthenticationConfig ReadFromJsonFile(string path)
+        {
+            IConfigurationRoot configuration;
+
+            var builder = new ConfigurationBuilder()
+             .SetBasePath(Directory.GetCurrentDirectory())
+            .AddJsonFile(path);
+
+            configuration = builder.Build();
+            return configuration.Get<AuthenticationConfig>();
+        }
+
+        public override string ToString()
+        {
+            return $"@Instance={this.Instance}\n@Tenant={this.Tenant}\n@ClientId={this.ClientId}\n@ClientSecret={this.ClientSecret}\n@CertificateName={this.CertificateName}\n@TodoListBaseAddress={this.TodoListBaseAddress}\n@TodoListScope={this.TodoListScope}";
+        }
+    }
+}
