+++ TestProject1/PrivateEndpointCallHelper.cs
@@ -1,0 +1,70 @@
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Net.Http;
+using System.Net.Http.Headers;
+using System.Threading.Tasks;
+
+namespace TestProject1
+{
+    /// <summary>
+    /// Helper class to call a protected API and process its result
+    /// </summary>
+    public class PrivateEndpointCallHelper
+    {
+        /// <summary>
+        /// Constructor
+        /// </summary>
+        /// <param name="httpClient">HttpClient used to call the protected API</param>
+        public PrivateEndpointCallHelper(HttpClient httpClient)
+        {
+            HttpClient = httpClient;
+        }
+
+        protected HttpClient HttpClient { get; private set; }
+
+        /// <summary>
+        /// Calls the protected web API and processes the result
+        /// </summary>
+        /// <param name="webApiUrl">URL of the web API to call (supposed to return Json)</param>
+        /// <param name="accessToken">Access token used as a bearer security token to call the web API</param>
+        /// <param name="processResult">Callback used to process the result of the call to the web API</param>
+        public async Task GetResultAsync(string webApiUrl, string accessToken, Action<IEnumerable<JObject>> processResult)
+        {
+            Trace.WriteLine("@CallPrivateEndpointAndProcessResultAsync()");
+
+            if (!string.IsNullOrEmpty(accessToken))
+            {
+                var defaultRequestHeaders = HttpClient.DefaultRequestHeaders;
+                if (defaultRequestHeaders.Accept == null || !defaultRequestHeaders.Accept.Any(m => m.MediaType == "application/json"))
+                {
+                    HttpClient.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
+                }
+                defaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", accessToken);
+
+                Trace.WriteLine("@PreGetAsync()");
+                HttpResponseMessage response = await HttpClient.GetAsync(webApiUrl);
+                Trace.WriteLine("@PostGetAsync()");
+
+                if (response.IsSuccessStatusCode)
+                {
+                    string json = await response.Content.ReadAsStringAsync();
+
+                    var result = JsonConvert.DeserializeObject<List<JObject>>(json);
+
+                    processResult(result);
+                }
+                else
+                {
+                    // Note that if you got reponse.Code == 403 and reponse.content.code == "Authorization_RequestDenied"
+                    // this is because the tenant admin as not granted consent for the application to call the Web API
+
+                    string content = await response.Content.ReadAsStringAsync();
+                }
+            }
+        }
+    }
+}
