+++ TestProject1/appsettings.json
@@ -1,9 +1,9 @@
 {
-	"Instance": "https://login.microsoftonline.com/{0}",
-	"Tenant": "[Enter here the tenantID or domain name for your Azure AD tenant]",
-	"ClientId": "[Enter here the ClientId for your application]",
-	"ClientSecret": "[Enter here a client secret for your application]",
-	"CertificateName": "[Or instead of client secret: Enter here the name of a certificate (from the user cert store) as registered with your application]",
-	"TodoListBaseAddress": "https://localhost:9999",
-	"TodoListScope": "api://[Enter_client_ID_Of_TodoListService-v2_from_Azure_Portal,_e.g._2ec40e65-ba09-4853-bcde-bcb60029e596]/.default"
+  "Instance": "https://login.microsoftonline.com/{0}",
+  "Tenant": "[Enter here the tenantID or domain name for your Azure AD tenant]",
+  "ClientId": "[Enter here the ClientId for your application]",
+  "ClientSecret": "[Enter here a client secret for your application]",
+  "CertificateName": "[Or instead of client secret: Enter here the name of a certificate (from the user cert store) as registered with your application]",
+  "TodoListBaseAddress": "https://localhost:9999",
+  "TodoListScope": "api://[Enter_client_ID_Of_TodoListService-v2_from_Azure_Portal,_e.g._2ec40e65-ba09-4853-bcde-bcb60029e596]/.default"
 }
