+++ WebApplication1/Properties/launchSettings.json
@@ -14,9 +14,9 @@
       "launchBrowser": true,
       "launchUrl": "swagger",
       "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
+        //"ASPNETCORE_ENVIRONMENT": "Development"
       },
-      "applicationUrl": "https://localhost:7107;http://localhost:5107",
+      "applicationUrl": "https://localhost:7107",
       "dotnetRunMessages": true
     },
     "IIS Express": {
@@ -24,7 +24,7 @@
       "launchBrowser": true,
       "launchUrl": "swagger",
       "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
+        //"ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
     "Docker": {
@@ -39,8 +39,8 @@
       "launchBrowser": true,
       "launchUrl": "https://localhost:7107/swagger",
       "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "ASPNETCORE_URLS": "https://localhost:7107;http://localhost:5107"
+        //"ASPNETCORE_ENVIRONMENT": "Development",
+        "ASPNETCORE_URLS": "https://localhost:7107"
       },
       "distributionName": ""
     }
