+++ WebApplication1/appsettings.json
@@ -1,18 +1,4 @@
 {
-/*
-The following identity settings need to be configured
-before the project can be successfully executed.
-For more info see https://aka.ms/dotnet-template-ms-identity-platform
-*/
-  "AzureAd": {
-    "Instance": "https://login.microsoftonline.com/",
-    "Domain": "qualified.domain.name",
-    "TenantId": "22222222-2222-2222-2222-222222222222",
-    "ClientId": "11111111-1111-1111-11111111111111111",
-
-    "Scopes": "access_as_user",
-    "CallbackPath": "/signin-oidc"
-  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
